requirements.txt
pandas==2.0.1
matplotlib==3.7.1
ast
seaborn==0.12.2
sklearn==1.3.1
pickle
-q -U google-generativeai
python-dotenv
flask-cors
werkzeug
numpy